añadido middleware email.js, iniciado email de confirmación con resend, añadida validación del token de acceso

+++ controllers/UsuariosController.js
@@ -1,13 +1,17 @@
+process.loadEnvFile()
 const Usuarios = require('../models/Usuarios')
+const { sendEmail, confirmEmail } = require('../middlewares/email')
 const {
   checkID,
   error500,
   checkData,
   checkEmailDuplicado,
+  checkTokenData,
 } = require('../middlewares/validations')
 
 const bcrypt = require('bcryptjs')
 const jwt = require('jsonwebtoken')
+const JWT_SECRET = process.env.JWT_SECRET
 
 const UsuariosController = {
   async getUsuarios(req, res) {
@@ -41,7 +45,33 @@ const UsuariosController = {
         return
       const Password = bcrypt.hashSync(req.body.Password, 10)
       const usuario = await Usuarios.create({ ...req.body, Password })
-      res.status(201).send({ message: 'usuario creado', usuario })
+      const idToken = jwt.sign({ _id: usuario._id }, process.env.JWT_SECRET)
+      // revisar notas de middleware email.js
+      // await sendEmail(req.body.Email, idToken)
+      res.status(201).send({
+        message: 'usuario creado, revisa tu email para confirmar el registro',
+        usuario,
+      })
+    } catch (error) {
+      if (error?.code === 11000)
+        return res
+          .status(401)
+          .send({ message: 'el email ya existe, elige otro' })
+
+      if (error?.errors?.Email?.kind === 'regexp')
+        return res.status(401).send({ message: 'formato de email erróneo' })
+      error500(error, res)
+    }
+  },
+
+  async confirmarUsuario(req, res) {
+    try {
+      const token = req.params.idToken
+      const payload = jwt.verify(token, JWT_SECRET)
+      await Usuarios.findByIdAndUpdate(payload._id, {
+        $set: { Confirmado: 'true' },
+      })
+      confirmEmail(res, 'usuarios')
     } catch (error) {
       error500(error, res)
     }
@@ -78,6 +108,11 @@ const UsuariosController = {
 
     try {
       if (checkID(usuarioId, res)) return
+      // revisamos que hay token en la request, y que coincida con el usuario
+      if (!req.headers.authorization)
+        return res.status(401).send({ message: 'falta el token' })
+      if (checkTokenData('Usuario', usuarioId, req.headers.authorization, res))
+        return
       const usuarioBorrado = await Usuarios.findByIdAndDelete(usuarioId)
       usuarioBorrado === null
         ? res.status(400).send({ message: 'id de usuario no encontrado' })
